CLI: Move the rest of the command to the CLI and add a prompt for the subscription email
