+++ packages/frontity/src/cli/index.ts
@@ -26,10 +26,13 @@ tsNode.register({
 });
 
 import program from "commander";
-import { dev, build, serve, subscribe, info, unknown } from "../commands";
+import { dev, build, serve } from "../commands";
 
 import { default as create } from "./create";
 import { default as createPackage } from "./create-package";
+import { default as subscribe } from "./subscribe";
+import { default as unknown } from "./unknown";
+import { default as info } from "./info";
 
 import packageJson from "../../package.json";
 
@@ -89,7 +92,7 @@ program
   .action(serve);
 
 program
-  .command("subscribe <email>")
+  .command("subscribe [email]")
   .description("Subscribe to Frontity newsletter.")
   .action(subscribe);
